+++ trunk/windows/HubFrame.h
@@ -368,7 +368,7 @@ private:
 	TStringList::size_type curCommandPosition;		//can't use an iterator because StringList is a vector, and vector iterators become invalid after resizing
 
 	tstring currentNeedle;		// search in chat window
-	int currentNeedlePos;		// search in chat window
+	long currentNeedlePos;		// search in chat window
 	void findText(tstring const& needle) throw();
 	tstring findTextPopup();
